refactor: split out virtual modules
